docker/python_train_models/requirements.txt
pandas==2.0.3
scikit-surprise==1.1.4
psycopg2-binary==2.9.10
mlflow==2.17.2
scikit-learn==1.3.2
numpy==1.24.4
joblib==1.2.0
sqlalchemy==1.4.31